refactor: moves mockCompanies from support to fixtures file
